[yee]  accordion삭제 및 파일합치고 정리(aceEditors, Modals에)
