+++ src/pages/Battle/components/AceEditors.js
@@ -13,27 +13,22 @@ import 'ace-builds/src-noconflict/theme-solarized_light';
 import 'ace-builds/src-noconflict/theme-terminal';
 import 'ace-builds/src-noconflict/ext-language_tools';
 
-const AceEditorPlayer = (props) => {
-    const { mode, theme } = props
-
+export const AceEditorPlayer = (props) => {
+    const { mode, theme } = props;
     function onChangeOne(newValue) {
         console.log('1:', newValue);
     }
-
     const JsDefault = `function solution(num) { 
     var answer = '';
         return answer;
 }`;
-
     const JavaDefault = `public String solution(int num) {
     String answer = '';
         return answer;
 }`
-
     const PythonDefault = `def solution(num):
         answer = ''
         return answer`;
-
     const DefaultTemp = "//함수와 변수를 임의로 변경하지 마세요" + `\n` + JavaDefault;
 
     return (
@@ -61,7 +56,54 @@ const AceEditorPlayer = (props) => {
                 placeholder="Placeholder Text"
             />
         </>
-    )
-}
+    );
+};
 
-export default AceEditorPlayer
+export const AceEditorOpp = (props) => {
+   const { mode, theme } = props;
+
+   function onChangeTwo(newValue) {
+      console.log('2:', newValue);
+   }
+
+   const JsDefault = `function solution(num) { 
+        var answer = '';
+           return answer;
+     }`;
+
+   const JavaDefault = `public String solution(int num) {
+    String answer = '';
+        return answer;
+}`;
+
+   const PythonDefault = `def solution(num):
+        answer = ''
+        return answer`;
+
+   const DefaultTempTwo =
+      '//함수와 변수를 임의로 변경하지 마세요' + `\n` + JavaDefault;
+   return (
+      <>
+         <AceEditor
+            width="100%"
+            height="100%"
+            style={{ borderRadius: '3px' }}
+            mode={mode}
+            theme={theme}
+            fontSize={15}
+            onChange={onChangeTwo}
+            editorProps={{ $blockScrolling: true }}
+            showPrintMargin={false}
+            readOnly={true}
+            highlightActiveLine={true}
+            setOptions={{
+               enableBasicAutocompletion: true,
+               enableLiveAutocompletion: true,
+               enableSnippets: true,
+            }}
+            value={DefaultTempTwo}
+            placeholder="Placeholder Text"
+         />
+      </>
+   );
+};
